python-backend/requirements.txt
Django>=4.2,<5.0
djangorestframework==3.15.2
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
django-extensions==3.2.3
psycopg[binary]
gunicorn
Pillow


